scripts/requirements.txt
requests
pdfminer.six 